Manually get `XDG_*_HOME` and drop pyxdg from deps
Drop `pyxdg` as dependency and instead get `XDG_CACHE_HOME` and
`XDG_CONFIG_HOME` manually.
The same way is being used as `pyxdg` does:
https://gitlab.freedesktop.org/xdg/pyxdg/-/blob/bd999c1c/xdg/BaseDirectory.py#L37-L44

There is a name clash between the two packages `xdg` and `pyxdg`. See:
https://github.com/srstevenson/xdg/issues/75
https://gitlab.freedesktop.org/xdg/pyxdg/-/issues/24

As only the two environment variables are being used from the library
they can just as good be expanded manually.
This solves the confusion and problems for users who have both
installed and it also reduces the dependency footprint.